+++ src/main/java/yapp/bestFriend/controller/ProductController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import yapp.bestFriend.model.dto.DefaultRes;
 import yapp.bestFriend.model.dto.request.CreateProductRequest;
 import yapp.bestFriend.model.dto.request.UpdateProductRequest;
-import yapp.bestFriend.model.dto.response.SimpleProductResponse;
+import yapp.bestFriend.model.dto.res.SimpleProductResponse;
 import yapp.bestFriend.model.utils.UserUtil;
 import yapp.bestFriend.service.ProductService;
